requirements.txt
streamlit
PyPDF2
python-docx
pandas
openpyxl
transformers
torch
sentencepiece
googletrans==4.0.0-rc1
langdetect